+++ src/scss/clientPortal/glenrockHay/styles.scss
@@ -18,7 +18,7 @@ body.style-guide {
     margin-top: 0;
 }
 
-.typography {
+.typography.glenrockHay {
     h1 {
         color: $green_dark_50;
         font-family: 'Sura', serif;
