+++ src/scss/clientPortal/wondaiCountryFestival/styles.scss
@@ -16,7 +16,7 @@ body.style-guide {
     margin-top: 0;
 }
 
-.typography {
+.typography.wondaiCountryFestival {
     h1 {
         color: $red_dark_50;
         font-family: 'Alegreya', serif;
